cleaing some rough edges and refactoring
