+++ src/components/SearchBar.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 
 class SearchBar extends Component {
   render() {
-    console.log("this.props.key", this.props.name);
+    
     return (
       <div className="col-12">
         <input
