+++ src/store/actions/authentication.js
@@ -30,7 +30,7 @@ export const checkForExpiredToken = () => {
       // Decode token and get user info
       const user = jwt_decode(token);
 
-      console.log((user.exp - currentTime) / 60);
+      
 
       // Check token expiration
       if (user.exp >= currentTime) {
