+++ src/store/actions/channels.js
@@ -14,9 +14,6 @@ export const getAllChannels = () => {
       let response = await instance.get("/channels/");
       let channels = response.data;
 
-      console.log(
-        "zerodebug => channels action => getAllChannels => res.data: ".channels
-      );
       dispatch({
         type: actionTypes.GET_ALL_CHANNELS,
         payload: channels
@@ -29,15 +26,15 @@ export const getAllChannels = () => {
 };
 
 export const getChannelMsgs = (chID, time = "") => {
-  console.log(chID);
+  
 
   return async dispatch => {
     try {
-      console.log("getChannelMsgs => ttime: ", time, chID);
+      
       let response = await instance.get(`/channels/${chID}/?latest=${time}`);
       let msgs = response.data;
 
-      console.log("getChannelMsgs => channel msgs: ", msgs);
+      
       dispatch({
         type: actionTypes.GET_CHANNEL_MSGS,
         payload: msgs,
@@ -60,15 +57,11 @@ export const getChannelInfo = chID => {
 export const postChannel = (newCh, history) => {
   return async dispatch => {
     try {
-      console.log("New channel obj: ", newCh);
+      
       let response = await instance.post("/channels/create/", newCh);
 
       let newChObj = response.data;
 
-      console.log(
-        "zerodebug => channels action => postChannel => res.data: ",
-        newChObj
-      );
       dispatch({
         type: actionTypes.POST_CHANNEL,
         payload: newChObj
@@ -89,11 +82,11 @@ export const postChannel = (newCh, history) => {
 };
 
 export const postMsg = (msg, chID) => {
-  console.log(msg);
+  
   return async dispatch => {
     try {
-      console.log("New msg sent: ", msg);
-      console.log("chID: ", chID);
+      
+      
 
       let response = await instance.post(`/channels/${chID}/send/`, msg);
 
@@ -103,8 +96,6 @@ export const postMsg = (msg, chID) => {
         type: actionTypes.POST_MSG,
         payload: newMsg
       });
-
-      
     } catch (error) {
       dispatch(setErrors(error));
       console.error(error.response.data);
